requirements.txt
python-dotenv
sentence-transformers
transformers
accelerate
pyodbc
numpy
pandas
requests
sqlalchemy
